Add changes to block modification for creator/lastModifer annotations
