+++ pkg/apis/serving/v1alpha1/service_lifecycle_test.go
@@ -16,11 +16,9 @@ limitations under the License.
 package v1alpha1
 
 import (
-	"context"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	authv1 "k8s.io/api/authentication/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"knative.dev/pkg/apis"
@@ -29,7 +27,6 @@ import (
 	duckv1beta1 "knative.dev/pkg/apis/duck/v1beta1"
 	apitesting "knative.dev/pkg/apis/testing"
 
-	"knative.dev/serving/pkg/apis/serving"
 	"knative.dev/serving/pkg/apis/serving/v1beta1"
 )
 
@@ -613,109 +610,4 @@ func TestServiceGetGroupVersionKind(t *testing.T) {
 	if got := s.GetGroupVersionKind(); got != want {
 		t.Errorf("got: %v, want: %v", got, want)
 	}
-}
-
-func TestAnnotateUserInfo(t *testing.T) {
-	const (
-		u1 = "oveja@knative.dev"
-		u2 = "cabra@knative.dev"
-		u3 = "vaca@knative.dev"
-	)
-
-	withUserAnns := func(u1, u2 string, s *Service) *Service {
-		a := s.GetAnnotations()
-		if a == nil {
-			a = map[string]string{}
-			s.SetAnnotations(a)
-		}
-		a[serving.CreatorAnnotation] = u1
-		a[serving.UpdaterAnnotation] = u2
-		return s
-	}
-
-	tests := []struct {
-		name     string
-		user     string
-		this     *Service
-		prev     *Service
-		wantAnns map[string]string
-	}{{
-		name: "create-new",
-		user: u1,
-		this: &Service{},
-		prev: nil,
-		wantAnns: map[string]string{
-			serving.CreatorAnnotation: u1,
-			serving.UpdaterAnnotation: u1,
-		},
-	}, {
-		// Old objects don't have the annotation, and unless there's a change in
-		// data they won't get it.
-		name:     "update-no-diff-old-object",
-		user:     u1,
-		this:     &Service{},
-		prev:     &Service{},
-		wantAnns: map[string]string{},
-	}, {
-		name: "update-no-diff-new-object",
-		user: u2,
-		this: withUserAnns(u1, u1, &Service{}),
-		prev: withUserAnns(u1, u1, &Service{}),
-		wantAnns: map[string]string{
-			serving.CreatorAnnotation: u1,
-			serving.UpdaterAnnotation: u1,
-		},
-	}, {
-		name: "update-diff-old-object",
-		user: u2,
-		this: &Service{
-			Spec: ServiceSpec{
-				DeprecatedRelease: &ReleaseType{},
-			},
-		},
-		prev: &Service{
-			Spec: ServiceSpec{
-				DeprecatedRunLatest: &RunLatestType{},
-			},
-		},
-		wantAnns: map[string]string{
-			serving.UpdaterAnnotation: u2,
-		},
-	}, {
-		name: "update-diff-new-object",
-		user: u3,
-		this: withUserAnns(u1, u2, &Service{
-			Spec: ServiceSpec{
-				DeprecatedRelease: &ReleaseType{},
-			},
-		}),
-		prev: withUserAnns(u1, u2, &Service{
-			Spec: ServiceSpec{
-				DeprecatedRunLatest: &RunLatestType{},
-			},
-		}),
-		wantAnns: map[string]string{
-			serving.CreatorAnnotation: u1,
-			serving.UpdaterAnnotation: u3,
-		},
-	}}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			ctx := apis.WithUserInfo(context.Background(), &authv1.UserInfo{
-				Username: test.user,
-			})
-			if test.prev != nil {
-				ctx = apis.WithinUpdate(ctx, test.prev)
-				test.prev.SetDefaults(ctx)
-			}
-			test.this.SetDefaults(ctx)
-			if got, want := test.this.GetAnnotations(), test.wantAnns; !cmp.Equal(got, want) {
-				t.Errorf("Annotations = %v, want: %v, diff (-got, +want): %s", got, want, cmp.Diff(got, want))
-			}
-		})
-	}
-
 }
